FixedPriceSignatureMinter review fixes (#164)
* Moves DOMAIN_SEPARATOR to FixedPriceSignatureMinter

* uint128 mintId in _editionMintData key

+++ contracts/modules/RangeEditionMinter.sol
@@ -22,7 +22,7 @@ contract RangeEditionMinter is IRangeEditionMinter, BaseMinter {
      * @dev Edition mint data
      * edition => mintId => EditionMintData
      */
-    mapping(address => mapping(uint256 => EditionMintData)) internal _editionMintData;
+    mapping(address => mapping(uint128 => EditionMintData)) internal _editionMintData;
 
     /**
      * @dev Number of tokens minted by each buyer address
